React costum hook - UserPerson
