+++ src/components/person-editor/usePerson.ts
@@ -1,0 +1,26 @@
+import {useState, useEffect} from 'react';
+import localforage from 'localforage';
+import type { Person } from '../../types/person';
+
+function saveperson(person:Person | null):void{
+    console.log("Saving person : " , person);
+    localforage.setItem("person", person);
+}
+
+export function usePerson(initialPerson:Person){
+    const [person, setPerson] = useState<Person | null>(null);
+
+    useEffect(() => {
+      const getPerson = async ()=> {
+          const person = await localforage.getItem<Person>("person")
+          setPerson(person ?? initialPerson);
+      }
+      getPerson();
+    },[initialPerson]);
+  
+    useEffect(() => {
+      saveperson(person);
+    },  [person]);
+
+    return [person, setPerson] as const;
+}
